More debuging code.

+++ make-release-tarball.bash
@@ -33,9 +33,15 @@ echo "Setting juju-core tree to $JUJU_CORE_BRANCH $REVNO."
 
 echo "Updating juju-core dependencies to the required versions."
 GOPATH=$WORK go get -v launchpad.net/godeps
-echo "$(ls $WORK)"
-echo "$(ls $WORK/bin/)"
-GODEPS=$WORK/bin/godeps
+GOPATH=$WORK go install -v launchpad.net/godeps
+GODEPS="$WORK/bin/godeps"
+if [[ ! -f $GODEPS ]]; then
+    echo "! Could not install godeps."
+    ls $WORK
+    ls $work/src/launchpad.net/godeps/
+    ls $WORK/bin/
+    exit 1
+di
 GOPATH=$WORK $GODEPS -u "${WORK}/src/launchpad.net/juju-core/dependencies.tsv"
 # Remove godeps.
 rm -r $WORK/bin
